Reorganizacao de codigo (separa domRenderers e eventListeners). Formulario da base ok. TODO: Criar renderizacao codicional de formulario e renderizacao por projeto
